Moved maintainer package under pkg/tbtc
