+++ pkg/tbtc/maintainer/bitcoin_difficulty.go
